+++ dao_universal_mongodb_test.go
@@ -4,92 +4,83 @@ import (
 	"fmt"
 	"math/rand"
 	"os"
-	"regexp"
 	"strconv"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/btnguyen2k/consu/reddo"
 	"github.com/btnguyen2k/godal"
-	"github.com/btnguyen2k/godal/sql"
 	"github.com/btnguyen2k/prom"
-	_ "github.com/mattn/go-sqlite3"
+	"go.mongodb.org/mongo-driver/bson"
 )
 
-func _cleanupSqlite(dir string) error {
-	return os.RemoveAll(dir)
+func _cleanupMongo(mc *prom.MongoConnect, collectionName string) error {
+	return mc.GetCollection(collectionName).Drop(nil)
 }
 
-func _testSqliteInitSqlConnect(t *testing.T, name string) *prom.SqlConnect {
-	dbDir := "./temp"
-	dbName := "tempdb"
-	if err := _cleanupSqlite(dbDir); err != nil {
-		t.Fatalf("%s/%s failed: %s", name, "_cleanupSqlite", err)
+func _testMongoInitMongoConnect(t *testing.T, testName, collectionName string) *prom.MongoConnect {
+	mongoUrl := strings.ReplaceAll(os.Getenv("MONGO_URL"), `"`, "")
+	if mongoUrl == "" {
+		t.Skipf("%s skipped", testName)
+		return nil
 	}
-	sqlc, err := NewSqliteConnection(dbDir, dbName, "sqlite3", 10000, nil)
+	mongoDb := strings.ReplaceAll(os.Getenv("MONGO_DB"), `"`, "")
+	if mongoDb == "" {
+		mongoDb = "test"
+	}
+	mc, err := NewMongoConnection(mongoUrl, mongoDb, 10000)
 	if err != nil {
-		t.Fatalf("%s/%s failed: %s", name, "NewSqliteConnection", err)
+		t.Fatalf("%s/%s failed: %s", testName, "NewMongoConnection", err)
 	}
-	return sqlc
+	if err := _cleanupMongo(mc, collectionName); err != nil {
+		t.Fatalf("%s/%s failed: %s", testName, "_cleanupMongo", err)
+	}
+	return mc
 }
 
-func TestNewSqliteConnection(t *testing.T) {
-	name := "TestNewSqliteConnection"
-	sqlc := _testSqliteInitSqlConnect(t, name)
-	defer sqlc.Close()
+func TestNewMongoConnection(t *testing.T) {
+	name := "TestNewMongoConnection"
+	mc := _testMongoInitMongoConnect(t, name, "table_temp")
+	defer mc.Close(nil)
 }
 
-func TestInitSqliteTable(t *testing.T) {
-	name := "TestInitSqliteTable"
-	sqlc := _testSqliteInitSqlConnect(t, name)
-	defer sqlc.Close()
-	tblName := "table_temp"
-	colDef := map[string]string{"col_email": "VARCHAR(64)", "col_age": "INT"}
+func TestInitMongoCollection(t *testing.T) {
+	name := "TestInitMongoCollection"
+	collectionName := "table_temp"
+	mc := _testMongoInitMongoConnect(t, name, collectionName)
+	defer mc.Close(nil)
 	for i := 0; i < 2; i++ {
-		if err := InitSqliteTable(sqlc, tblName, colDef); err != nil {
+		if err := InitMongoCollection(mc, collectionName); err != nil {
 			t.Fatalf("%s failed: %s", name, err)
 		}
 	}
 }
 
-func TestCreateIndexSqlite(t *testing.T) {
-	name := "TestCreateIndexSqlite"
-	sqlc := _testSqliteInitSqlConnect(t, name)
-	defer sqlc.Close()
-	tblName := "table_temp"
-	colDef := map[string]string{"col_email": "VARCHAR(64)", "col_age": "INT"}
-	if err := InitSqliteTable(sqlc, tblName, colDef); err != nil {
+func _testMongoInit(t *testing.T, name, collectionName string) (*prom.MongoConnect, UniversalDao) {
+	mc := _testMongoInitMongoConnect(t, name, collectionName)
+	if err := InitMongoCollection(mc, collectionName); err != nil {
 		t.Fatalf("%s failed: %s", name, err)
 	}
-	if err := CreateIndexSql(sqlc, tblName, true, []string{"col_email"}); err != nil {
-		t.Fatalf("%s failed: %s", name, err)
+	index := map[string]interface{}{
+		"key":    map[string]interface{}{"email": 1},
+		"name":   "uidx_email",
+		"unique": true,
 	}
-	if err := CreateIndexSql(sqlc, tblName, false, []string{"col_age"}); err != nil {
-		t.Fatalf("%s failed: %s", name, err)
+	mc.CreateCollectionIndexes(collectionName, []interface{}{index})
+	mongoUrl := strings.ReplaceAll(os.Getenv("MONGO_URL"), `"`, "")
+	txModeOnWrite := strings.Contains(mongoUrl, "replicaSet=")
+	if txModeOnWrite {
+		fmt.Println("txModeOnWrite:", txModeOnWrite)
 	}
+	return mc, NewUniversalDaoMongo(mc, collectionName, txModeOnWrite)
 }
 
-func _testSqliteInit(t *testing.T, name, tblName string) (*prom.SqlConnect, UniversalDao) {
-	sqlc := _testSqliteInitSqlConnect(t, name)
-	colDef := map[string]string{"col_email": "VARCHAR(64)", "col_age": "INT"}
-	if err := InitSqliteTable(sqlc, tblName, colDef); err != nil {
-		t.Fatalf("%s failed: %s", name, err)
-	}
-	if err := CreateIndexSql(sqlc, tblName, true, []string{"col_email"}); err != nil {
-		t.Fatalf("%s failed: %s", name, err)
-	}
-	if err := CreateIndexSql(sqlc, tblName, false, []string{"col_age"}); err != nil {
-		t.Fatalf("%s failed: %s", name, err)
-	}
-	extraColNameToFieldMappings := map[string]string{"col_email": "email", "col_age": "age"}
-	return sqlc, NewUniversalDaoSql(sqlc, tblName, true, extraColNameToFieldMappings)
-}
-
-func TestSqlite_Create(t *testing.T) {
-	name := "TestSqlite_Create"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_Create(t *testing.T) {
+	name := "TestMongo_Create"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo := NewUniversalBo("id", 1357)
 	ubo.SetDataAttr("name.first", "Thanh")
 	ubo.SetDataAttr("name.last", "Nguyen")
@@ -103,22 +94,11 @@ func TestSqlite_Create(t *testing.T) {
 	}
 }
 
-func _isErrorDuplicatedEntrySqlite(err error) bool {
-	if err == godal.GdaoErrorDuplicatedEntry {
-		return true
-	}
-	errStr := fmt.Sprintf("%e", err)
-	// fmt.Printf("%#e\n", err)
-	return regexp.MustCompile(`\WErrNo=19\W`).FindString(errStr) != "" &&
-		(regexp.MustCompile(`\WErrNoExtended=1555\W`).FindString(errStr) != "" ||
-			regexp.MustCompile(`\WErrNoExtended=2067\W`).FindString(errStr) != "")
-}
-
-func TestSqlite_CreateExistingPK(t *testing.T) {
-	name := "TestSqlite_CreateExistingPK"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_CreateExistingPK(t *testing.T) {
+	name := "TestMongo_CreateExistingPK"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo := NewUniversalBo("id", 1357)
 	ubo.SetDataAttr("name.first", "Thanh")
 	ubo.SetDataAttr("name.last", "Nguyen")
@@ -132,18 +112,18 @@ func TestSqlite_CreateExistingPK(t *testing.T) {
 	}
 
 	ubo.SetExtraAttr("email", "myname2@mydomain.com")
-	if ok, err := dao.Create(ubo); !_isErrorDuplicatedEntrySqlite(err) {
+	if ok, err := dao.Create(ubo); err != godal.GdaoErrorDuplicatedEntry {
 		t.Fatalf("%s failed: %s", name, err)
 	} else if ok {
 		t.Fatalf("%s failed: record should not be created twice", name)
 	}
 }
 
-func TestSqlite_CreateExistingUnique(t *testing.T) {
-	name := "TestSqlite_CreateExistingUnique"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_CreateExistingUnique(t *testing.T) {
+	name := "TestMongo_CreateExistingUnique"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo := NewUniversalBo("id", 1357)
 	ubo.SetDataAttr("name.first", "Thanh")
 	ubo.SetDataAttr("name.last", "Nguyen")
@@ -157,18 +137,18 @@ func TestSqlite_CreateExistingUnique(t *testing.T) {
 	}
 
 	ubo.SetId("id2")
-	if ok, err := dao.Create(ubo); !_isErrorDuplicatedEntrySqlite(err) {
+	if ok, err := dao.Create(ubo); err != godal.GdaoErrorDuplicatedEntry {
 		t.Fatalf("%s failed: %s", name, err)
 	} else if ok {
 		t.Fatalf("%s failed: record should not be created twice", name)
 	}
 }
 
-func TestSqlite_CreateGet(t *testing.T) {
-	name := "TestSqlite_CreateGet"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_CreateGet(t *testing.T) {
+	name := "TestMongo_CreateGet"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo := NewUniversalBo("id", 1357)
 	ubo.SetDataAttr("name.first", "Thanh")
 	ubo.SetDataAttr("name.last", "Nguyen")
@@ -204,11 +184,11 @@ func TestSqlite_CreateGet(t *testing.T) {
 	}
 }
 
-func TestSqlite_CreateDelete(t *testing.T) {
-	name := "TestSqlite_CreateDelete"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_CreateDelete(t *testing.T) {
+	name := "TestMongo_CreateDelete"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo := NewUniversalBo("id", 1357)
 	ubo.SetDataAttr("name.first", "Thanh")
 	ubo.SetDataAttr("name.last", "Nguyen")
@@ -240,11 +220,11 @@ func TestSqlite_CreateDelete(t *testing.T) {
 	}
 }
 
-func TestSqlite_CreateGetMany(t *testing.T) {
-	name := "TestSqlite_CreateGetMany"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_CreateGetMany(t *testing.T) {
+	name := "TestMongo_CreateGetMany"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 
 	idList := make([]string, 0)
 	for i := 0; i < 10; i++ {
@@ -272,11 +252,11 @@ func TestSqlite_CreateGetMany(t *testing.T) {
 	}
 }
 
-func TestSqlite_CreateGetManyWithFilter(t *testing.T) {
-	name := "TestSqlite_CreateGetManyWithFilter"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_CreateGetManyWithFilter(t *testing.T) {
+	name := "TestMongo_CreateGetManyWithFilter"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 
 	idList := make([]string, 0)
 	for i := 0; i < 10; i++ {
@@ -297,7 +277,7 @@ func TestSqlite_CreateGetManyWithFilter(t *testing.T) {
 		}
 	}
 
-	filter := &sql.FilterFieldValue{Field: "col_age", Operation: ">=", Value: 35 + 3}
+	filter := bson.M{"age": bson.M{"$gte": 35 + 3}}
 	if boList, err := dao.GetAll(filter, nil); err != nil {
 		t.Fatalf("%s failed: %s", name, err)
 	} else if len(boList) != 7 {
@@ -305,11 +285,11 @@ func TestSqlite_CreateGetManyWithFilter(t *testing.T) {
 	}
 }
 
-func TestSqlite_CreateGetManyWithSorting(t *testing.T) {
-	name := "TestSqlite_CreateGetManyWithSorting"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_CreateGetManyWithSorting(t *testing.T) {
+	name := "TestMongo_CreateGetManyWithSorting"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 
 	idList := make([]string, 0)
 	for i := 0; i < 10; i++ {
@@ -330,7 +310,7 @@ func TestSqlite_CreateGetManyWithSorting(t *testing.T) {
 		}
 	}
 
-	sorting := map[string]string{"col_email": "desc"}
+	sorting := bson.M{"email": -1}
 	if boList, err := dao.GetAll(nil, sorting); err != nil {
 		t.Fatalf("%s failed: %s", name, err)
 	} else {
@@ -342,11 +322,11 @@ func TestSqlite_CreateGetManyWithSorting(t *testing.T) {
 	}
 }
 
-func TestSqlite_CreateGetManyWithFilterAndSorting(t *testing.T) {
-	name := "TestSqlite_CreateGetManyWithFilterAndSorting"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_CreateGetManyWithFilterAndSorting(t *testing.T) {
+	name := "TestMongo_CreateGetManyWithFilterAndSorting"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 
 	idList := make([]string, 0)
 	for i := 0; i < 10; i++ {
@@ -367,8 +347,8 @@ func TestSqlite_CreateGetManyWithFilterAndSorting(t *testing.T) {
 		}
 	}
 
-	filter := &sql.FilterFieldValue{Field: "col_email", Operation: "<", Value: "3@mydomain.com"}
-	sorting := map[string]string{"col_email": "desc"}
+	filter := bson.M{"email": bson.M{"$lt": "3@mydomain.com"}}
+	sorting := bson.M{"email": -1}
 	if boList, err := dao.GetAll(filter, sorting); err != nil {
 		t.Fatalf("%s failed: %s", name, err)
 	} else if len(boList) != 3 {
@@ -380,53 +360,52 @@ func TestSqlite_CreateGetManyWithFilterAndSorting(t *testing.T) {
 	}
 }
 
-// // currently godal does not support SQLite flavor!
-// func TestSqlite_CreateGetManyWithSortingAndPaging(t *testing.T) {
-// 	name := "TestSqlite_CreateGetManyWithSortingAndPaging"
-// 	tblName := "table_temp"
-// 	sqlc, dao := _testSqliteInit(t, name, tblName)
-// 	defer sqlc.Close()
-//
-// 	idList := make([]string, 0)
-// 	for i := 0; i < 10; i++ {
-// 		idList = append(idList, strconv.Itoa(i))
-// 	}
-// 	rand.Seed(time.Now().UnixNano())
-// 	rand.Shuffle(len(idList), func(i, j int) { idList[i], idList[j] = idList[j], idList[i] })
-// 	for i := 0; i < 10; i++ {
-// 		ubo := NewUniversalBo(idList[i], uint64(i))
-// 		ubo.SetDataAttr("name.first", strconv.Itoa(i))
-// 		ubo.SetDataAttr("name.last", "Nguyen")
-// 		ubo.SetExtraAttr("email", idList[i]+"@mydomain.com")
-// 		ubo.SetExtraAttr("age", 35+i)
-// 		if ok, err := dao.Create(ubo); err != nil {
-// 			t.Fatalf("%s failed: %s", name, err)
-// 		} else if !ok {
-// 			t.Fatalf("%s failed: cannot create record", name)
-// 		}
-// 	}
-//
-// 	fromOffset := 3
-// 	numRows := 4
-// 	sorting := map[string]string{"col_email": "desc"}
-// 	if boList, err := dao.GetN(fromOffset, numRows, nil, sorting); err != nil {
-// 		t.Fatalf("%s failed: %s", name, err)
-// 	} else if len(boList) != numRows {
-// 		t.Fatalf("%s failed: expected %#v items but received %#v", name, numRows, len(boList))
-// 	} else {
-// 		for i := 0; i < numRows; i++ {
-// 			if boList[i].GetId() != strconv.Itoa(9-i-fromOffset) {
-// 				t.Fatalf("%s failed: expected record %#v but received %#v", name, strconv.Itoa(9-i-fromOffset), boList[i].GetId())
-// 			}
-// 		}
-// 	}
-// }
+func TestMongo_CreateGetManyWithSortingAndPaging(t *testing.T) {
+	name := "TestMongo_CreateGetManyWithSortingAndPaging"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 
-func TestSqlite_Update(t *testing.T) {
-	name := "TestSqlite_Update"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+	idList := make([]string, 0)
+	for i := 0; i < 10; i++ {
+		idList = append(idList, strconv.Itoa(i))
+	}
+	rand.Seed(time.Now().UnixNano())
+	rand.Shuffle(len(idList), func(i, j int) { idList[i], idList[j] = idList[j], idList[i] })
+	for i := 0; i < 10; i++ {
+		ubo := NewUniversalBo(idList[i], uint64(i))
+		ubo.SetDataAttr("name.first", strconv.Itoa(i))
+		ubo.SetDataAttr("name.last", "Nguyen")
+		ubo.SetExtraAttr("email", idList[i]+"@mydomain.com")
+		ubo.SetExtraAttr("age", 35+i)
+		if ok, err := dao.Create(ubo); err != nil {
+			t.Fatalf("%s failed: %s", name, err)
+		} else if !ok {
+			t.Fatalf("%s failed: cannot create record", name)
+		}
+	}
+
+	fromOffset := 3
+	numRows := 4
+	sorting := bson.M{"email": -1}
+	if boList, err := dao.GetN(fromOffset, numRows, nil, sorting); err != nil {
+		t.Fatalf("%s failed: %s", name, err)
+	} else if len(boList) != numRows {
+		t.Fatalf("%s failed: expected %#v items but received %#v", name, numRows, len(boList))
+	} else {
+		for i := 0; i < numRows; i++ {
+			if boList[i].GetId() != strconv.Itoa(9-i-fromOffset) {
+				t.Fatalf("%s failed: expected record %#v but received %#v", name, strconv.Itoa(9-i-fromOffset), boList[i].GetId())
+			}
+		}
+	}
+}
+
+func TestMongo_Update(t *testing.T) {
+	name := "TestMongo_Update"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo := NewUniversalBo("id", 1357)
 	ubo.SetDataAttr("name.first", "Thanh")
 	ubo.SetDataAttr("name.last", "Nguyen")
@@ -470,11 +449,11 @@ func TestSqlite_Update(t *testing.T) {
 	}
 }
 
-func TestSqlite_UpdateNotExist(t *testing.T) {
-	name := "TestSqlite_UpdateNotExist"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_UpdateNotExist(t *testing.T) {
+	name := "TestMongo_UpdateNotExist"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo := NewUniversalBo("id", 1357)
 	ubo.SetDataAttr("name.first", "Thanh")
 	ubo.SetDataAttr("name.last", "Nguyen")
@@ -488,11 +467,11 @@ func TestSqlite_UpdateNotExist(t *testing.T) {
 	}
 }
 
-func TestSqlite_UpdateDuplicated(t *testing.T) {
-	name := "TestSqlite_UpdateDuplicated"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_UpdateDuplicated(t *testing.T) {
+	name := "TestMongo_UpdateDuplicated"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 
 	ubo1 := NewUniversalBo("1", 1357)
 	ubo1.SetDataAttr("name.first", "Thanh")
@@ -512,16 +491,16 @@ func TestSqlite_UpdateDuplicated(t *testing.T) {
 	}
 
 	ubo1.SetExtraAttr("email", "2@mydomain.com")
-	if _, err := dao.Update(ubo1); !_isErrorDuplicatedEntrySqlite(err) {
+	if _, err := dao.Update(ubo1); err != godal.GdaoErrorDuplicatedEntry {
 		t.Fatalf("%s failed: %s", name, err)
 	}
 }
 
-func TestSqlite_SaveNew(t *testing.T) {
-	name := "TestSqlite_SaveNew"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_SaveNew(t *testing.T) {
+	name := "TestMongo_SaveNew"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo := NewUniversalBo("id", 1357)
 	ubo.SetDataAttr("name.first", "Thanh")
 	ubo.SetDataAttr("name.last", "Nguyen")
@@ -559,11 +538,11 @@ func TestSqlite_SaveNew(t *testing.T) {
 	}
 }
 
-func TestSqlite_SaveExisting(t *testing.T) {
-	name := "TestSqlite_SaveExisting"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_SaveExisting(t *testing.T) {
+	name := "TestMongo_SaveExisting"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo := NewUniversalBo("id", 1357)
 	ubo.SetDataAttr("name.first", "Thanh")
 	ubo.SetDataAttr("name.last", "Nguyen")
@@ -625,11 +604,11 @@ func TestSqlite_SaveExisting(t *testing.T) {
 	}
 }
 
-func TestSqlite_SaveExistingUnique(t *testing.T) {
-	name := "TestSqlite_SaveExistingUnique"
-	tblName := "table_temp"
-	sqlc, dao := _testSqliteInit(t, name, tblName)
-	defer sqlc.Close()
+func TestMongo_SaveExistingUnique(t *testing.T) {
+	name := "TestMongo_SaveExistingUnique"
+	collectionName := "table_temp"
+	mc, dao := _testMongoInit(t, name, collectionName)
+	defer mc.Close(nil)
 	ubo1 := NewUniversalBo("1", 1357)
 	ubo1.SetDataAttr("name.first", "Thanh1")
 	ubo1.SetDataAttr("name.last", "Nguyen1")
@@ -648,7 +627,7 @@ func TestSqlite_SaveExistingUnique(t *testing.T) {
 	}
 
 	ubo1.SetExtraAttr("email", "2@mydomain.com")
-	if _, _, err := dao.Save(ubo1); !_isErrorDuplicatedEntrySqlite(err) {
+	if _, _, err := dao.Save(ubo1); err != godal.GdaoErrorDuplicatedEntry {
 		t.Fatalf("%s failed: %s", name, err)
 	}
 }